Progress in SPA
