+++ components/layout/Layout.js
@@ -1,4 +1,4 @@
-import Header from './Header';
+import Header from '../Header';
 import Footer from '../Footer';
 import LayoutWithSidebar from './LayoutWithSidebar';
