+++ package.json
@@ -17,7 +17,6 @@
     },
     "dependencies": {
         "@apollo/react-hooks": "^3.1.5",
-        "@loadable/component": "^5.12.0",
         "@zeit/next-css": "^1.0.1",
         "apollo-boost": "^0.4.9",
         "babel-plugin-styled-components": "^1.10.7",
